Fix detecting changes upon update properly
**Why?**

CodeCommit has a limit of 100 files per commit.

To workaround this limit, the implementation walked across all the files
and split it into chunks of 99 files. However, when one of those chunks would
not have any changes, it would raise an exception that was handled at the
incorrect level. Causing it to abort the creation of the PR all together.
Even when there were changes in another chunk.

**What?**

Fixed handling the "no changes in this chunk" error at the chunk level.
So it will only skip that chunk and continue with the next.

Maintaining a list of commits that were created, and if any were, it will open
the pull request.

Additionally, the initial commit code had a lot of repetitive code.
This got refactored as well.

Since there are two initial_commit.py functions, the changes had to be applied
twice.