Await SFN executions before bootstrapping continues
**Why?**

As described in issue #518, the bootstrap pipeline fails to perform the main.py
code when the account creation or bootstrapping process is still in progress.

**What?**

The code changes ensure the script will wait for any Step Function executions
that are triggered by the sync_to_s3.py process. It will wait for 30 seconds in
a loop until they succeeded.