Abort bootstrap pipeline when SFN error occurred
**Why?**

As the account management and bootstrapping steps are performed in Step
Function State Machines, the errors might not be noticed until a follow-up error
occurs when trying to interact with one of the failing accounts.

**What?**

Modified the bootstrap pipeline to check if these state machines do not have
any aborted, timed out, or failed executions. If they do, it will log the error
and instruct the user to look into the fault first.