Fix CloudFormation permissions to protect and delete stacks and tag IAM
**Why?**

The policy in the pipeline management attached to the CloudFormation role
does not allow it to DeleteStack or UpdateTerminationProtection.
As those were using the RequestTag condition that is not compatible
with these API calls.

Additionally, the `iam:TagResource` action does not exist.

**What?**

For the `DeleteStack` and `UpdateTerminationProtection` actions, it is
updated to use the `iam:ResourceTag` condition instead.

While the `iam:TagResource` action got replaced with `iam:TagPolicy` and
`iam:TagRole` actions.