Fix pipeline generation policies (#533)
* Reverting Pipeline Generation restrictions

**Why?**

With a recent PR, an attempt was made to restrict the `iam:PassRole`
permissions for the pipeline generator. However, since ADF supports configuring
which role to use, this would introduce a breaking change.

**What?**

Reverting to the previous policy of `iam:PassRole`.

Plan is to address this in the next major ADF release.

* Fix CloudFormation permissions to protect and delete stacks and tag IAM

**Why?**

The policy in the pipeline management attached to the CloudFormation role
does not allow it to DeleteStack or UpdateTerminationProtection.
As those were using the RequestTag condition that is not compatible
with these API calls.

Additionally, the `iam:TagResource` action does not exist.

**What?**

For the `DeleteStack` and `UpdateTerminationProtection` actions, it is
updated to use the `iam:ResourceTag` condition instead.

While the `iam:TagResource` action got replaced with `iam:TagPolicy` and
`iam:TagRole` actions.