Delete pipeline regions SSM Param after successful CFN delete operation
**Why?**

At the moment, when it determines which pipelines need to be deleted, it will
also delete the parameters.

However, if the deletion fails, the parameters have been deleted and it will
not be able to detect that the stack needs to be deleted still.

**What?**

Moved the deletion of the SSM Parameters to the pipeline deletion state
machine.