Fix account creation wait for bootstrap to complete (#537)
* Forward CodePipeline Execution Id to Account Mgmt Creation SFN

Step 2 of fixing #518.

**Why?**

As explained in #518, we need to forward the execution id of the CodePipeline
that triggered the Account Management state machine so we can wait for to
complete.

**What?**

Adding the CodePipeline execution identifier to the Step Functions
State Machine invocation to enable tracing state machines in progress at
CodeBuild execution time.

* Only process account and deployment maps generated by same ADF version

**Why?**

When an account file or deployment map is updated, it should only be processed
when the version equals the current ADF version. Otherwise it should skip the
file.

This will ensure we don't get into compatibility issues, where a file structure
update will make the processing of the file fail.

**What?**

Checking the version number attached to the S3 object metadata against the
current ADF version number. If those mismatch, it will skip the file.

* Fix Doc to use arguments

* Limit SFN Execution id to 80 chars

**Why?**

SFN has a limit of 80 chars.

* Await SFN executions before bootstrapping continues

**Why?**

As described in issue #518, the bootstrap pipeline fails to perform the main.py
code when the account creation or bootstrapping process is still in progress.

**What?**

The code changes ensure the script will wait for any Step Function executions
that are triggered by the sync_to_s3.py process. It will wait for 30 seconds in
a loop until they succeeded.

* Abort bootstrap pipeline when SFN error occurred

**Why?**

As the account management and bootstrapping steps are performed in Step
Function State Machines, the errors might not be noticed until a follow-up error
occurs when trying to interact with one of the failing accounts.

**What?**

Modified the bootstrap pipeline to check if these state machines do not have
any aborted, timed out, or failed executions. If they do, it will log the error
and instruct the user to look into the fault first.

* Use CodePipeline execution id instead of CodeBuild for sync_to_s3 ops

* Install helper requirements as part of the bootstrap pipeline

* Catch key not found in sync_to_s3 helper

**Why?**

The implementation of the head_object throws a low-level
`botocore.exceptions.ClientError` instead of
the higher level `s3_client.exceptions.NoSuchKey`.

**What?**

Properly caught the error using the low-level approach.