Fix account bootstrap on OU move
**Why?**

When you move an account from the root to an OU, an event is emitted that should
trigger the Account Bootstrap state machine.

However, the task role that was configured did not have the necessary
permissions to invoke the state machine.

**What?**

A new role is created that will enable the event to trigger the Account
Bootstrap state machine.

While investigating the issue, it was also noticed that the original task role
had permissions that were no longer required. Those stale permissions are
removed too.