Fix initial commit on fresh install (#544)
**Why?**

When the repository is created, there are no commits in the repo yet.
The process tried to create a branch where the parent commit id was empty.
This is not allowed.

**What?**

Instead, it should only create the branch when there is a commit to branch off
from. Otherwise it should only create the new commits on the main branch
directly.