Fix storing ADF config before it needs to be read the first time
**Why?**

After deploying ADF for the first time, it will need to store the adfconfig.yml
values in the AWS Systems Manager Parameter Store. Such that the logic inside
the account bootstrap and account management state machines can execute
correctly.

At the moment, the config is written after the Step Functions are triggered by
syncing account configurations to S3. This will fail on the read of the
`config` parameter.

**What?**

Adding another step that will only store the configuration values in the ADF
Bootstrap CodeBuild execution. Before it copies the files that would trigger
the account management and account bootstrapping state machines.