Trigger pipeline regeneration upon organization changes
**Why?**

As described in issue #549. When an account is moved in or out of an
organization unit (OU), the pipelines that include that OU are not updated
accordingly.

**What?**

By checking the type of trigger that started this CodePipeline execution, it
will force the upload process of the deployment map files when the pipeline
was triggered using a StartPipelineExecution API Call. As these are used
when the pipeline is instructed to start by a Lambda function that detected an
organization unit change.

* Add support for --force to the `sync_to_s3.py` helper.
* Add `describe_codepipeline_trigger.py` helper to fetch the trigger type of a
  specific execution.
* Set the `--force` as `$EXTRA_OPTS` that are passed to the `sync_to_s3.py`
  helper if needed.
* Allow `aws-deployment-framework-pipelines` CodeBuild container to fetch the
  `aws-deployment-framework-pipelines` executions.