Fix generate_params.py to rely on pipeline input definitions instead
**Why?**

The `generate_params.py` script relied on the `account_ous.json` file that was
uploaded to the shared modules bucket.

While ADF got support for newer account target selection mechanisms, the
`account_ous.json` file structure was not updated accordingly.

To ensure future support and ease to maintain new account target selections
it was decided to drop the `account_ous.json` and move to the pipeline input
definition instead.

**What?**

* Updated generate_params.py to rely on the pipeline input definition instead.
* Updated generate_params.py to ensure that parameters would be generated
  correctly. Before it was using a complex structure inside the resolver class.
  With the new changes, it relies on generating a new copy of the parameters
  with each step instead. Making it easier to test and harder to break.
* Updated the resolver class to a simplified version using easy to extend
  resolver classes.
* Added multiple test cases to ensure correctness of the generate_params.py
  script.