Fix readability of pipeline generation executions by Step Functions
**Why?**

At the moment, when a pipeline definition is created or updated, a Lambda
function triggers the pipeline generation state machine.

While triggering this state machine, it does not specify a name for the
execution. So Step Function will generate its own unique name instead.

However, when you look at the pipeline generation state machine, it is
impossible to tell which generation is running or has failed from the list of
executions. As all executions have unique ids.

**What?**

Changed the start execution process when a pipeline is created or updated, such
that it will include the pipeline name at the start of the execution.

Followed by the CodePipeline execution id of the
aws-deployment-framework-pipelines process that created or modified the
deployment map that contained this pipeline.

As well as the Lambda execution id, such that rerunning the lambda function will
also result in unique execution ids to pass to the Step Function.