Fix parameter validation upon install/update (#554)
* Fix parameter validation upon install/update

**Why?**

When installing or updating ADF, it did not validate the input given by
the user. In some situations, this implied that it would attempt to
install/update failing minutes later. Or it might not work at all after
the CloudFormation deployment of the SAR template reports back as if
all went well.

One of the parameters, the deployment region was particularly dangerous
to leave blank on an update. As that would change the bucket name that
would be used by ADF to share resources.

**What?**

Added validations to the template.

The account alias regex was found here [0].
While the role name requirements were found at [1].

* [0](https://docs.aws.amazon.com/IAM/latest/APIReference/API_CreateAccountAlias.html)
* [1](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-iam-role.html#cfn-iam-role-rolename)

* Improve regex validation according to review

The Account Name requirements were fount at [0].

* [0](https://docs.aws.amazon.com/organizations/latest/APIReference/API_Account.html)