Add support for CodeStar CodeBuild clone ref and update CodeStar setup
**Why?**

The CodeStar provider did not support the `CODEBUILD_CLONE_REF` artifact mode
yet. This is useful if you would like to access the git repository, for example
to analyze the history or checkout a different branch.

For CodeStar Connections, it was trying to resolve the CodeStar Connection Arn
from SSM Parameter Store at the pipeline stack generation. However, resolving
parameters should be performed in the input generation step.

**What?**

* Added support to define the `output_artifact_format` in the CodeStar source
  provider. Allowing `CODEBUILD_CLONE_REF` to be defined as the output format.
* Moved resolving the `codestar_connection_path` in SSM Parameter Store to the
  pipeline input generation. As this is where it should be performed.
  The Lambda function resolving these parameters has the required permissions
  already.
* Added support for our CloudFormation role to perform
  `codestar-connections:PassConnection`, such that CodeStar connections can be
  configured in the CodePipeline stack that it generates.