Improve adf-pipelines CodeBuild permissions kickstarting SFN
**Why?**

With the move to Step Functions, the majority of the workload for the
adf-pipelines CodeBuild process was moved to the ADF Pipeline Management State
Machine in AWS Step Functions.

However, the policies attached to the original CodeBuild environment were not
updated yet. Additionally, the machine type used for this process could be
reduced in size too.

**What?**

* Changed the name from `pipeline-provisioner*` to
  `pipeline-generation-provisioner*`, as it no longer privisions the pipelines
  but ensures the Step Function has the required input to do it instead.
* Removed the original provisioner role from AssumeRole permissions that it no
  longer requires.
* Removed environment variables that are not used by the process to provision
  the pipeline generator.
* Removed policy statements that the pipeline-generation provisioner does not
  require.
* Applied the `/adf-automation/` path to the pipeline-generation provisioner
  role.