Add CodeBuild VPC permission to default permissions to provision pipelines
**Why?**

With the new pipeline management Step Function, the role used to deploy the
pipeline CloudFormation stacks changed. This role is not exposed and I think
we should not change that to keep the pipeline generation process in ADF its
control.

However, this would not allow anyone to modify the permissions required to
use CodeBuild in a specific VPC.

**What?**

* To keep it simple, the CodeBuild VPC permissions were moved into the default
  CloudFormation roles that deploy the pipelines.
* Removed the CodeBuild VPC permissions docs, as those steps are no longer
  required.