Fix permissions to allow CodeBuild as a deployment provider
**Why?**

The CloudFormation role that is used to deploy the pipelines was only allowing
CodeBuild projects which name started with `adf-build-`.
However, it should also allow `adf-deploy-` resources, as that is how the
CodeBuild project is named when it is used as a deployment provider.

**What?**

Added the `adf-deploy-*` project name for CodeBuild projects deployed by the
pipeline generation process.