Improve account cross account access management
**Why?**

At the moment, with each region, it would perform the PutRolePolicy and
GetRolePolicy calls required to update the target and deployment accounts.

On top of that, it would update the deployment account roles with every account
it updated too.

**What?**

Reducing the number of API calls by applying the deployment account role
policies only when we update the deployment account.

Additionally, we update the target accounts after iterating over the regions
and determining the full list of S3 buckets and KMS keys to grant access to.