Merge in default deployment props into target specifications
**Why?**

The default provider, properties, and regions were not merged into
the target configuration yet.

This moved the lookup of the defaults to the execution step of the
pipeline deployment.

If we move applying the defaults to the input stage, this makes it easier
to see what will be applied at execution level. As that input is captured
in the new Pipeline Management state machine.

**What?**

* Apply default deploy provider configuration to the target configuration at
  input generation time.