Fix generate params regions lookup + refactor pipeline management (#584)
* Fix generate params regions lookup

**Why?**

With the changes introduced in PR #525, support was added for different
deployment map providers.

This changed the location of the pipeline parameters to the scope of the
source. In the case of our S3 deployment map provider, that changed:
`/deployment/${pipeline_name}/regions` to
`/deployment/S3/${pipeline_name}/regions`.

In the `generate_params.py` helper, it still continued to look for the regions
in the original SSM Parameter Store location.

**What?**

* Refactored pipeline management to separate execution details from the
  pipeline definition, the pipeline input, and the SSM parameters retrieved.
  * Renaming `input` to `pipeline_input` to clearly indicate what input we are
    referring to.
* Refactored line length issues spotted while making changes to forward the
  deployment map source and name.
* Added the `deployment_map_source` and `deployment_map_name` context to the
  ADF Pipeline Management state machine.
* Fixed an issue in the pipeline generation, where the EventBridge rule would
  be defined if the account_id would be set when CodeCommit was not used as
  the source provider.
* Removed
  `${bootstrap_repo}/adf-build/shared/cdk/generate_pipeline_inputs.py` as that
  is replaced by the ADF Pipeline Management using the Step Function state
  machine instead.
* Moved flattening lists to the list_utils helper, making it easier to reuse
  this logic elsewhere. Since the lists had to be sorted and made unique each
  time this logic was moved to this function too. Reducing repeated code and
  different variations trying to achieve the same thing.

* Support debug logs in generate_params.py

**Why?**

We need more debug information to enable debugging generate params efficiently.

**What?**

* Added debug log levels.
* End-users can add the `ADF_LOG_LEVEL` environment variable and set it to
  `DEBUG` to get the logs printed.

* Merge in default deployment props into target specifications

**Why?**

The default provider, properties, and regions were not merged into
the target configuration yet.

This moved the lookup of the defaults to the execution step of the
pipeline deployment.

If we move applying the defaults to the input stage, this makes it easier
to see what will be applied at execution level. As that input is captured
in the new Pipeline Management state machine.

**What?**

* Apply default deploy provider configuration to the target configuration at
  input generation time.

* S3 Deploy Extract should use the default False when not set