Update AWS SAM CLI to resolve OpenSSL dep issue + upgrade others
**Why?**

Recently, cryptography dropped support for OpenSSL 1.1.0 and older.
This dependency is included via aws-sam-cli. Unfortunately this was not
version pinned, causing the `generate_params.py` dependency installation to
break all of a sudden.

Luckily version 1.68 of aws-sam-cli fixed this issue.

A few requirements.txt files were not included in our tox setup.
This made it harder to validate that all dependencies link to the same version
correctly.

**What?**

* Update to aws-sam-cli 1.69.
* Update to aws-cdk 1.188.0
* Update other dependencies to the latest compatible version too.
* Added missing requirements.txt files in the tox.ini file, to ensure we don't
  miss out on updating these too. As it will break/warn if there are
  incompatible version requirements.
* Sorted requirements alphabetically where they were not.