Fix bootstrap mirror pipelines
**Why?**

An error was generated when the `aws-deployment-framework-bootstrap` is
mirrored from another location, via a pipeline that is deployed with ADF.

In this case, the `generate_params.py` file will look for the new
`pipeline_input` attribute in the pipeline definition. While the pipeline
definition was generated with a previous version of ADF, thereby using the
`input` attribute key instead.

**What?**

Introduced backward compatibility to the `input` key if the `pipeline_input`
key is missing in the pipeline definition.

Also added a comment that this support for backwards compatibility will be
dropped in the next major release of ADF.