Fix CodeDeploy sample to work in gov-cloud
**Why?**

At the moment, the install script for CodeDeploy would install CodeDeploy using
the global S3 URL to fetch the file from the S3 bucket.

As such:
```
https://aws-codedeploy-${REGION}.s3.amazonaws.com/latest/install
```

However, the CodeDeploy Gov cloud buckets are not available through this
global URL. So when the REGION=`us-gov-east-1` or `us-gov-west-1` it would
fail.

**What?**

To fix this, we should use the regional URL instead:

```
https://aws-codedeploy-${REGION}.s3.${REGION}.amazonaws.com/latest/install
```