Upgrade to latest CDK + SAM CLI, Python 3.10 (Py3.11 in CodeBuild)
**Why?**

As this commit will update the new default CodeBuild container image to use,
we will pick the most recent Standard CodeBuild image that is available at the
time of this commit. Which is STANDARD_7_0.

This allows us to continue to support the version for the foreseeable future.
As we rather not need to change this frequently, as that would introduce a
breaking change.