Improve docs and add CodeStar Connection policy (#649)
**Why?**

There are a few requirements to adopt CodeStar connections that were not
documented yet. This change request addresses those and introduces the required
policy as an example in the example-global-iam.yml template.

**What?**

CodeStar Connections are the preferred method to link with GitHub nowadays.
Hence, updated the admin guide to highlight this preference.