Remove CodeCommitRole from initial base stack (#663)
The adf-codecommit-role-base role from the ADF CloudFormation Initial Base Stack for the Management Account in the us-east-1 region, with the trust policy being assumable from `codecommit.amazonaws.com` doesn't seem be used anywhere. I couldn't find it being used in ADF codebase, or the cloudformation template it comes in. I couldn't anything find you could do with codecommit that requires a role with `codecommit.amazonaws.com` being the trusted service principal to assume to it or even passing a role to codecommit.