Fix initial value type of allow-empty-targets
**Why?**

ADF fails to build in the bootstrap pipeline if no `allow-empty-targets` value
is configured in the `adfconfig.yml` file.

Root cause, the default value is a boolean False. While the Parameter Store
service expects values to be of type string.

**What?**

This change set fixes it by converting the default to string when written to
Parameter Store.