Auto generate unique branch names on new version deployments
**Why?**

In case a previous deployment of ADF failed for whatever reason, running it
again would require you to delete the branch it created the first time.

**What?**

Since we don't want to end up deleting a branch that we did not create, the
proposed solution is to generate a unique name for the branch and use that.

1. It will start with the version number as the branch name, like `v3.2.0`.
2. If that exists, it will iterate from `v3.2.0-no-1` to `v3.2.0-no-9`.
3. Highly unlikely, but if all of these exist, it will fallback
   to `v3.2.0-at-{unix_timestamp}`.