feat: add multi-org generate_params.py (#672)
## Why?

When running ADF in a multi-org setup, this feature allows you to set stage-specific files in the repositories' `params` folder. This feature doesn't break existing parameter generation. It just adds a new feature.

## Example

Let's assume that we have a three AWS Org setup with a dev, int and prod AWS Org. 
That implies that the SSM param "/adf/org/stage" will have one of the following values: [dev, int, prod]. For now, let's assume the value is set to "dev".
Let's further assume that, the `params` folder of our target repository has the following structure:

- params:
   - global_dev.yml
   - global_int.yml
   - global_prod.yml
   - global.yml

With this feature, the param file global_dev.yml will be selected based on the value "dev" in SSM param "/adf/org/stage".

---------

Co-authored-by: Andreas Falkenberg
Co-authored-by: Simon Kok