Fix support for pipelines running CodeBuild containers with Python < v3.10
**Why?**

With the current dependency requirements, a build pipeline will fail to install
liburl3 if a Python version prior to v3.10 is used.

As botocore defined a Python version specific requirement for urllib3, as can
be seen at https://github.com/boto/botocore/blob/1.34.17/setup.py#L28-L29.

**What?**

If urllib3 is installed in a CodeBuild container which has Python version
< v3.10, we should use the latest 1.26 release that is available instead.